Move LabelInfo and HLabelInfo to otx.core.types.label
