+++ tests/unit/core/conftest.py
@@ -10,8 +10,6 @@ from datumaro.components.annotation import AnnotationType, LabelCategories
 from datumaro.components.dataset import Dataset, DatasetItem
 from datumaro.components.media import Image
 from otx.core.config import register_configs
-from otx.core.data.dataset.base import LabelInfo
-from otx.core.data.dataset.classification import HLabelInfo
 from otx.core.data.entity.base import ImageInfo, Points
 from otx.core.data.entity.visual_prompting import (
     VisualPromptingBatchDataEntity,
@@ -21,6 +19,7 @@ from otx.core.data.entity.visual_prompting import (
     ZeroShotVisualPromptingBatchPredEntity,
     ZeroShotVisualPromptingDataEntity,
 )
+from otx.core.types.label import HLabelInfo, LabelInfo
 from torchvision import tv_tensors
