+++ tests/unit/core/metrics/test_accuracy.py
@@ -5,14 +5,13 @@
 
 import pytest
 import torch
-from otx.core.data.dataset.base import LabelInfo
-from otx.core.data.dataset.classification import HLabelInfo
 from otx.core.metrics.accuracy import (
     HlabelAccuracy,
     MixedHLabelAccuracy,
     MulticlassAccuracywithLabelGroup,
     MultilabelAccuracywithLabelGroup,
 )
+from otx.core.types.label import HLabelInfo, LabelInfo
 
 
 class TestAccuracy:
